Move RenderBoxHelper to separate file
